feat: refactored tool results - introduced a sealed interface to enable to collectively reference all tool result data classes, similar to a union type
