+++ adapter/src/main/kotlin/de/fraunhofer/iem/spha/adapter/tools/trivy/TrivyAdapter.kt
@@ -11,12 +11,12 @@ package de.fraunhofer.iem.spha.adapter.tools.trivy
 
 import de.fraunhofer.iem.spha.adapter.AdapterResult
 import de.fraunhofer.iem.spha.adapter.kpis.cve.CveAdapter
-import de.fraunhofer.iem.spha.model.adapter.trivy.CVSSData
-import de.fraunhofer.iem.spha.model.adapter.trivy.TrivyDto
-import de.fraunhofer.iem.spha.model.adapter.trivy.TrivyDtoV1
-import de.fraunhofer.iem.spha.model.adapter.trivy.TrivyDtoV2
-import de.fraunhofer.iem.spha.model.adapter.trivy.TrivyVulnerabilityDto
-import de.fraunhofer.iem.spha.model.adapter.vulnerability.VulnerabilityDto
+import de.fraunhofer.iem.spha.model.adapter.CVSSData
+import de.fraunhofer.iem.spha.model.adapter.TrivyDto
+import de.fraunhofer.iem.spha.model.adapter.TrivyDtoV1
+import de.fraunhofer.iem.spha.model.adapter.TrivyDtoV2
+import de.fraunhofer.iem.spha.model.adapter.TrivyVulnerabilityDto
+import de.fraunhofer.iem.spha.model.adapter.VulnerabilityDto
 import io.github.oshai.kotlinlogging.KotlinLogging
 import java.io.InputStream
 import kotlin.math.max
