+++ adapter/src/test/kotlin/de/fraunhofer/iem/spha/adapter/kpis/cve/CveAdapterTest.kt
@@ -11,7 +11,7 @@ package de.fraunhofer.iem.spha.adapter.kpis.cve
 
 import de.fraunhofer.iem.spha.adapter.AdapterResult
 import de.fraunhofer.iem.spha.adapter.ErrorType
-import de.fraunhofer.iem.spha.model.adapter.vulnerability.VulnerabilityDto
+import de.fraunhofer.iem.spha.model.adapter.VulnerabilityDto
 import kotlin.test.fail
 import org.junit.jupiter.api.Test
