+++ adapter/src/test/kotlin/de/fraunhofer/iem/spha/adapter/tools/tlc/TlcAdapterTest.kt
@@ -12,15 +12,15 @@ package de.fraunhofer.iem.spha.adapter.tools.tlc
 import de.fraunhofer.iem.spha.adapter.AdapterResult
 import de.fraunhofer.iem.spha.adapter.tools.tlc.model.ArtifactVersion
 import de.fraunhofer.iem.spha.adapter.tools.tlc.model.Version
-import de.fraunhofer.iem.spha.model.adapter.tlc.ArtifactDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.ArtifactVersionDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.DependencyGraphDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.DependencyNodeDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.EnvironmentInfoDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.ProjectDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.RepositoryInfoDto
-import de.fraunhofer.iem.spha.model.adapter.tlc.ScopeToGraph
-import de.fraunhofer.iem.spha.model.adapter.tlc.TlcDto
+import de.fraunhofer.iem.spha.model.adapter.ArtifactDto
+import de.fraunhofer.iem.spha.model.adapter.ArtifactVersionDto
+import de.fraunhofer.iem.spha.model.adapter.DependencyGraphDto
+import de.fraunhofer.iem.spha.model.adapter.DependencyNodeDto
+import de.fraunhofer.iem.spha.model.adapter.EnvironmentInfoDto
+import de.fraunhofer.iem.spha.model.adapter.ProjectDto
+import de.fraunhofer.iem.spha.model.adapter.RepositoryInfoDto
+import de.fraunhofer.iem.spha.model.adapter.ScopeToGraph
+import de.fraunhofer.iem.spha.model.adapter.TlcDto
 import de.fraunhofer.iem.spha.model.kpi.KpiType
 import kotlin.test.Test
 import kotlin.test.assertEquals
