+++ model/src/main/kotlin/de/fraunhofer/iem/spha/model/adapter/OsvVulnerabilityDto.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Fraunhofer IEM. All rights reserved.
+ * Copyright (c) 2024-2025 Fraunhofer IEM. All rights reserved.
  *
  * Licensed under the MIT license. See LICENSE file in the project root for details.
  *
@@ -7,7 +7,7 @@
  * License-Filename: LICENSE
  */
 
-package de.fraunhofer.iem.spha.model.adapter.osv
+package de.fraunhofer.iem.spha.model.adapter
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -34,7 +34,7 @@ data class Event(
 @Serializable
 data class Affected(
     @SerialName("package") val packageX: Package,
-    @SerialName("ranges") val ranges: List<Range>,
+    @SerialName("ranges") val osvVulnerabilityRanges: List<OsvVulnerabilityRange>,
 )
 
 @Serializable
@@ -62,7 +62,8 @@ data class Package(
 )
 
 @Serializable
-data class Range(
+@SerialName("range")
+data class OsvVulnerabilityRange(
     @SerialName("events") val events: List<Event>,
     @SerialName("repo") val repo: String,
     @SerialName("type") val type: String,
