+++ model/src/main/kotlin/de/fraunhofer/iem/spha/model/adapter/ToolResult.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2025 Fraunhofer IEM. All rights reserved.
+ *
+ * Licensed under the MIT license. See LICENSE file in the project root for details.
+ *
+ * SPDX-License-Identifier: MIT
+ * License-Filename: LICENSE
+ */
+
+package de.fraunhofer.iem.spha.model.adapter
+
+sealed interface ToolResult
