+++ model/src/main/kotlin/de/fraunhofer/iem/spha/model/adapter/TrivyDto.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Fraunhofer IEM. All rights reserved.
+ * Copyright (c) 2024-2025 Fraunhofer IEM. All rights reserved.
  *
  * Licensed under the MIT license. See LICENSE file in the project root for details.
  *
@@ -7,9 +7,8 @@
  * License-Filename: LICENSE
  */
 
-package de.fraunhofer.iem.spha.model.adapter.trivy
+package de.fraunhofer.iem.spha.model.adapter
 
-import de.fraunhofer.iem.spha.model.adapter.vulnerability.VulnerabilityDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonObject
@@ -19,13 +18,13 @@ data class TrivyDto(val vulnerabilities: Collection<VulnerabilityDto>)
 @Serializable
 data class TrivyDtoV1(
     @SerialName("Vulnerabilities") val vulnerabilities: List<TrivyVulnerabilityDto> = listOf()
-)
+) : ToolResult
 
 @Serializable
 data class TrivyDtoV2(
     @SerialName("Results") val results: List<Result> = listOf(),
     @SerialName("SchemaVersion") val schemaVersion: Int,
-)
+) : ToolResult
 
 @Serializable
 data class Result(
